+++ R/board_url.R
@@ -131,6 +131,8 @@ pin_browse.pins_board_url <- function(board, name, version = NULL, ..., cache = 
   }
 }
 
+
+
 # Unsupported features ----------------------------------------------------
 
 #' @export